requirements.txt
httpx==0.27.0
seleniumbase==4.28.5
selectolax==0.3.21
python-dotenv==1.0.1